websocket模块解耦 数字计算使用BigDecimal 代码整理

+++ java/src/main/java/com/toesbieya/my/model/entity/BizDocumentSub.java
@@ -2,11 +2,13 @@ package com.toesbieya.my.model.entity;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class BizDocumentSub {
     private Integer id;
     private String pid;
     private Integer cid;
     private String cname;
-    private Double num;
+    private BigDecimal num;
 }
