+++ java/src/main/java/com/toesbieya/my/model/entity/BizPurchaseOrder.java
@@ -4,6 +4,8 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
 
+import java.math.BigDecimal;
+
 @Data
 @ToString(callSuper = true)
 @EqualsAndHashCode(callSuper = true)
@@ -12,5 +14,5 @@ public class BizPurchaseOrder extends BizDocument<BizPurchaseOrderSub> {
     private String sname;
     private Integer finish;
     private Long ftime;
-    private Double total;
+    private BigDecimal total;
 }
