+++ java/src/main/java/com/toesbieya/my/model/vo/export/StockExport.java
@@ -5,6 +5,8 @@ import com.alibaba.excel.annotation.ExcelProperty;
 import com.alibaba.excel.annotation.write.style.ColumnWidth;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class StockExport {
     @ExcelIgnore
@@ -20,11 +22,11 @@ public class StockExport {
 
     @ExcelProperty(value = "库存总数", index = 2)
     @ColumnWidth(20)
-    private double total_num;
+    private BigDecimal total_num;
 
     @ExcelProperty(value = "库存总值", index = 3)
     @ColumnWidth(20)
-    private double total_price;
+    private BigDecimal total_price;
 
     @ExcelProperty(value = "采购订单号", index = 4)
     @ColumnWidth(20)
@@ -32,11 +34,11 @@ public class StockExport {
 
     @ExcelProperty(value = "采购单价", index = 5)
     @ColumnWidth(20)
-    private double cg_price;
+    private BigDecimal cg_price;
 
     @ExcelProperty(value = "采购数量", index = 6)
     @ColumnWidth(20)
-    private double cg_num;
+    private BigDecimal cg_num;
 
     @ExcelProperty(value = "采购入库单号", index = 7)
     @ColumnWidth(20)
@@ -48,5 +50,5 @@ public class StockExport {
 
     @ExcelProperty(value = "入库数量", index = 9)
     @ColumnWidth(20)
-    private double rk_num;
+    private BigDecimal rk_num;
 }
