+++ java/src/main/java/com/toesbieya/my/model/vo/result/StockSearchResult.java
@@ -2,10 +2,12 @@ package com.toesbieya.my.model.vo.result;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class StockSearchResult {
     private Integer cid;
     private String cname;
-    private Double total_num;
-    private Double total_price;
+    private BigDecimal total_num;
+    private BigDecimal total_price;
 }
