+++ java/src/main/java/com/toesbieya/my/model/vo/statictis/FourBlockStat.java
@@ -2,10 +2,12 @@ package com.toesbieya.my.model.vo.statictis;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class FourBlockStat {
-    private int online;
-    private double purchase;
-    private double sell;
-    private double profit;
+    private long online;
+    private BigDecimal purchase;
+    private BigDecimal sell;
+    private BigDecimal profit;
 }
